pyproject.toml
[project]
name = "markadoros"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jim Downie", email = "jd42@sanger.ac.uk" }
]
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "click>=8.3.1",
    "jsonschema>=4.26.0",
    "pandas>=2.3.3",
    "pymmseqs>=1.0.5",
    "pysam>=0.23.3",
    "scikit-learn>=1.8.0",
]

[project.scripts]
markadoros = "markadoros:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/markadoros"]
include = [
    "/src/markadoros/data/bold.json",
    "/src/markadoros/data/unite.json",
]

[build-system]
build-backend  = 'hatchling.build'
requires       = ['hatchling']
